requirements.txt
streamlit
langchain
langchain-ollama
selenium
beautifulsoup4
lxml
html5lib
python-dotenv
requests